Extract integration tests
